Remove old implementation of media queries
